+++ src/main/resources/application-test.properties
@@ -20,6 +20,6 @@ logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
 
 spring.data.mongodb.host=localhost
 spring.data.mongodb.port=27017
-spring.data.mongodb.database=epam
+spring.data.mongodb.database=epamtest
  spring.main.allow-bean-definition-overriding=true
 spring.main.allow-circular-references=true
