+++ src/test/java/com/epam/gymapp/integration/runners/TrainingComponentTest.java
@@ -1,4 +1,4 @@
-package com.epam.gymapp.integration;
+package com.epam.gymapp.integration.runners;
 
 import org.junit.runner.RunWith;
 
@@ -8,8 +8,8 @@ import io.cucumber.junit.CucumberOptions;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(
-    features = "src/test/resources/features",
-    glue = {"com.epam.gymapp.integration"},
+    features = "src/test/resources/features/integration",
+    glue = {"com.epam.gymapp.integration.steps"},
     plugin = {"pretty"}
 )
 public class TrainingComponentTest {
