+++ src/test/resources/features/integration/training_publish.feature
